+++ app/src/main/java/com/dps0340/attman/AttendancestatusActivity.java
@@ -1,17 +1,13 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
-import java.util.Timer;
-import java.util.TimerTask;
 
 import androidx.appcompat.app.AppCompatActivity;
