+++ app/src/main/java/com/dps0340/attman/CameraActivity.java
@@ -1,4 +1,4 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.os.Bundle;
