+++ app/src/main/java/com/dps0340/attman/HomeActivity.java
@@ -1,13 +1,10 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.content.Intent;
-import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.Bundle;
-import android.text.Html;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
