+++ app/src/main/java/com/dps0340/attman/LoginActivity.java
@@ -1,12 +1,10 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.Manifest;
-import android.app.Activity;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.os.Build;
 import android.os.Bundle;
-import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
