+++ app/src/main/java/com/dps0340/attman/MyView.java
@@ -1,21 +1,13 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
+import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.graphics.Color;
-import android.graphics.Typeface;
-import android.os.Bundle;
-import android.content.Context;
 import android.graphics.Canvas;
+import android.graphics.Color;
 import android.graphics.Paint;
-import android.view.MotionEvent;
+import android.graphics.Typeface;
 import android.view.View;
-import android.widget.ImageView;
-
-import com.android.volley.toolbox.Volley;
 
 class MyView extends View {
     public MyView(Context context) {
@@ -45,23 +37,4 @@ class MyView extends View {
         canvas.drawBitmap(sb2, 280, 1200, null);
     }
 
-}
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(new MyView(this));
-        //setContentView(R.layout.activity_main);
-    }
-
-    public boolean onTouchEvent(MotionEvent event) { //터치시 다음화면인 로그인화면으로 이동
-        int action = event.getAction();
-        if (action == MotionEvent.ACTION_DOWN) {
-            Intent intent = new Intent(MainActivity.this, LoginActivity.class);
-            startActivity(intent);
-        }
-        return super.onTouchEvent(event);
-    }
 }
