+++ app/src/main/java/com/dps0340/attman/PrivacyActivity.java
@@ -1,15 +1,8 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.TextView;
-import android.widget.Toast;
-
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.toolbox.Volley;
-import org.json.JSONException;
-import org.json.JSONObject;
 
 import androidx.appcompat.app.AppCompatActivity;
