+++ app/src/main/java/com/dps0340/attman/QrcodescanActivity.java
@@ -1,10 +1,9 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.TextView;
