+++ app/src/main/java/com/dps0340/attman/RegisterRequest.java
@@ -1,17 +1,15 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
-import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.Map;
 
 import com.android.volley.AuthFailureError;
-import com.android.volley.Request;
 import com.android.volley.Response;
 import com.android.volley.toolbox.StringRequest;
 
 public class RegisterRequest extends StringRequest {
     final static private String URL = "http://an950120.dothome.co.kr/Register.php";
-    private Map<String,String>map;
+    private final Map<String,String>map;
 
     public RegisterRequest(String userID, String userPassword, String userName, int userNumber, String userEmail,Response.Listener<String> listener) {
         super(Method.POST, URL, listener, null);
