+++ app/src/main/java/com/dps0340/attman/ResultActivity.java
@@ -1,11 +1,10 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
