+++ app/src/main/java/com/dps0340/attman/SelfdiagnosisActivity.java
@@ -1,10 +1,8 @@
-package com.example.attendance_management;
+package com.dps0340.attman;
 
 import android.content.ActivityNotFoundException;
-import android.content.Context;
 import android.content.Intent;
 import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
 import android.provider.MediaStore;
 import android.util.Pair;
@@ -13,21 +11,11 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
-import android.view.MotionEvent;
 import android.widget.Toast;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.ListIterator;
-
-import javax.xml.transform.Result;
 
 public class SelfdiagnosisActivity extends AppCompatActivity {
     private ArrayList<Pair<String, String>> symptomsList;
